% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koRpus.lang.pt-package.R
\docType{package}
\name{koRpus.lang.pt-package}
\alias{koRpus.lang.pt-package}
\title{The koRpus.lang.pt Package}
\description{
Language Support for 'koRpus' Package: Portuguese.
}
\details{
\tabular{ll}{
Package: \tab koRpus.lang.pt\cr
Type: \tab Package\cr
Version: \tab 0.1-3\cr
Date: \tab 2018-02-21\cr
Depends: \tab R (>= 3.1),koRpus (>= 0.11-2)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://reaktanz.de/?c=hacking&s=koRpus\cr
}

Adds support for the Portuguese language to the 'koRpus' package. To ask for help, report
bugs, suggest feature improvements, or discuss the global development of the package,
please consider subscribing to the koRpus-dev mailing list (<http://korpusml.reaktanz.de>).
}
\author{
Meik Michalke, with contributions from Katharina Karges, Jan Vanhove
}
\keyword{package}
