% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koRpus.lang.pt-package.R
\docType{package}
\name{koRpus.lang.pt-package}
\alias{koRpus.lang.pt-package}
\title{The koRpus.lang.pt Package}
\description{
Language support for koRpus: Portuguese.
}
\details{
\tabular{ll}{
Package: \tab koRpus.lang.pt\cr
Type: \tab Package\cr
Version: \tab 0.01-1\cr
Date: \tab 2016-07-05\cr
Depends: \tab R (>= 2.10.0),koRpus (>= 0.07-1), methods\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://reaktanz.de/?c=hacking&s=koRpus\cr
}

Adds support for the Portuguese language to the koRpus package. Due to some restrictions on CRAN,
      the full package sources are only available from the project homepage. To ask for help,
      report bugs, suggest feature improvements,
      or discuss the global development of the package,
      please subscribe to the koRpus-dev mailing list: https://ml06.ispgateway.de/mailman/listinfo/korpus-dev_r.reaktanz.de
}
\author{
m.eik michalke, with contributions from Katharina Karges, Jan Vanhove
}
\keyword{package}

