% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.support-pt.R
\name{lang.support.pt}
\alias{lang.support.pt}
\title{Language support for Portuguese}
\usage{
lang.support.pt()
}
\description{
This function adds support for Portuguese to the koRpus package. You should not
need to call it manually, as that is done automatically when this package is
loaded.
}
\details{
In particular, this function adds the following:
\itemize{
 \item \code{lang}: The additional language "pt" to be used with koRpus
 \item \code{hyphen}: An additional set of hyphenation patterns (see \code{\link{hyph.pt}})
 \item \code{treetag}: The additional preset "pt-utf8",
      implemented according to the respective
   TreeTagger[1] script
 \item \code{POS tags}: An additional set of tags,
      implemented using the documentation for the corresponding
   TreeTagger parameter set[2]
}
}
\references{
[1] \url{http://www.cis.uni-muenchen.de/~schmid/tools/TreeTagger/}

[2] \url{http://www.cis.uni-muenchen.de/~schmid/tools/TreeTagger/data/Portuguese-Tagset.html}
}

